Fix behavior when deleting unspecified terms

+++ app/models/term.rb
@@ -6,6 +6,11 @@ class Term < ActiveRecord::Base
       job.term = Term.find_or_create_by(label: "Unspecified")
       job.save
     end
-    self.destroy
+    if self.label == "Unspecified" && self.jobs.count != 0
+      self.errors.add(:base, "You must delete or reassign the jobs with unspecified terms.")
+      false
+    else
+      self.destroy
+    end
   end
 end
